+++ tests/testthat/test-garnish_affinity_assemble_generate.R
@@ -10,7 +10,7 @@ testthat::test_that("garnish_affinity_assemble_generate", {
      .[, MHC := c("HLA-A*02:01 HLA-DRB1*14:67")]
 
   # run test
-    dto <- garnish_affinity(dt, predict = FALSE, blast = FALSE, fitness = FALSE)
+    dto <- garnish_affinity(dt, predict = FALSE, blast = FALSE)
 
   testthat::expect_equal(dto %>% nrow, 183)
   testthat::expect_equal(dto %>% length, 78)