+++ tests/testthat/test-garnish_dissimilarity.R
@@ -2,9 +2,9 @@ testthat::test_that("garnish_dissimilarity", {
 
   v <- c("SIINFEKL")
 
-  st <- v %>% garnish_dissimilarity(db = "human")
+  st <- v %>% garnish_dissimilarity(db = "human", aval = 52)
 
-  testthat::expect_equal(st[, dissimilarity] %>% is.na,
-                        TRUE)
+  testthat::expect_equal(st[, dissimilarity] %>% format(digits = 5),
+                        c("1"))
 
 })