+++ tests/testthat/test-garnish_slim.R
@@ -13,14 +13,12 @@ testthat::test_that("garnish_slim", {
 
   testthat::expect_equal(max(dto$iedb_score, na.rm = TRUE), 5.94051e-05)
 
-  testthat::expect_equal(round(max(dto$NeoantigenRecognitionPotential, na.rm = TRUE), 4), 2.3688)
-
-  testthat::expect_equal(length(names(dto)), 20)
+  testthat::expect_equal(length(names(dto)), 18)
 
   testthat::expect_equal(nrow(dto[antigen.garnish_class != "Unclassified"]), 3)
 
   dto <- garnish_slim(dt, slimmer = FALSE)
 
-  testthat::expect_equal(length(names(dto)), 25)
+  testthat::expect_equal(length(names(dto)), 23)
 
 })