+++ tests/testthat/test-garnish_summary.R
@@ -21,8 +21,6 @@ testthat::test_that("garnish_summary", {
     testthat::expect_equal(dto2$nmers, 269)
     testthat::expect_equal(dto2$MHC_binders_class_I, 6)
     testthat::expect_equal(dto2$variants, 4)
-    testthat::expect_equal(ncol(dto2), 28)
-    testthat::expect_equal(dto2$garnish_score %>% signif(digits = 3), 74900)
-    testthat::expect_equal(dto2$fitness_scores_class_I %>% signif(digits = 3), 16.2)
+    testthat::expect_equal(ncol(dto2), 23)
 
     })